+++ src/main/java/org/jabref/gui/fieldeditors/EditorTextArea.java
@@ -6,10 +6,14 @@ import java.util.Objects;
 import java.util.ResourceBundle;
 import java.util.function.Supplier;
 
+import javafx.event.EventHandler;
 import javafx.fxml.Initializable;
+import javafx.scene.Node;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.TextArea;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
 
 import org.jabref.gui.ClipBoardManager;
 import org.jabref.gui.fieldeditors.contextmenu.EditorContextAction;
@@ -27,15 +31,19 @@ public class EditorTextArea extends TextArea implements Initializable, ContextMe
 
     public EditorTextArea() {
         this("");
+        System.out.println("Creating EditorTextArea with empty text");
     }
 
     public EditorTextArea(final String text) {
         super(text);
-
+        System.out.println("Creating EditorTextArea with text:");
         // Hide horizontal scrollbar and always wrap text
         setWrapText(true);
 
         ClipBoardManager.addX11Support(this);
+
+        // Add our custom Tab key event handler to traverse focus when empty.
+        addEventFilter(KeyEvent.KEY_PRESSED, new FieldTraversalEventHandler());
     }
 
     @Override
@@ -53,7 +61,7 @@ public class EditorTextArea extends TextArea implements Initializable, ContextMe
     }
 
     /**
-     * Set pasteActionHandler variable to passed handler
+     * Set pasteActionHandler variable to passed handler.
      *
      * @param handler an instance of PasteActionHandler that describes paste behavior
      */
@@ -63,11 +71,32 @@ public class EditorTextArea extends TextArea implements Initializable, ContextMe
     }
 
     /**
-     * Override javafx TextArea method applying TextArea.paste() and pasteActionHandler after
+     * Override JavaFX TextArea paste method to perform paste and then execute pasteActionHandler.
      */
     @Override
     public void paste() {
         super.paste();
         pasteActionHandler.run();
+    }
+
+    // Custom event handler for Tab key presses.
+    private static class FieldTraversalEventHandler implements EventHandler<KeyEvent> {
+        @Override
+        public void handle(KeyEvent event) {
+            if (event.getCode() == KeyCode.TAB && !event.isShiftDown() && !event.isControlDown()) {
+                event.consume();
+
+                // Get the current text area
+                Node node = (Node) event.getSource();
+                KeyEvent newEvent = new KeyEvent(node,
+                        event.getTarget(), event.getEventType(),
+                        event.getCharacter(), event.getText(),
+                        event.getCode(), event.isShiftDown(),
+                        true, event.isAltDown(),
+                        event.isMetaDown());
+
+                node.fireEvent(newEvent);
+            }
+        }
     }
 }